Moved the module from lib/twimock.rb to lib/twimock/api.rb
