+++ lib/twimock.rb
@@ -6,7 +6,7 @@ require "twimock/application"
 require "twimock/user"
 require "twimock/auth_hash"
 require "twimock/errors"
-require "sham_rack"
+require "twimock/api"
 
 module Twimock
   extend self
@@ -28,57 +28,5 @@ module Twimock
     else
       Twimock::AuthHash.new
     end 
-  end
-
-  module Api
-    extend self
-
-    API_HOST_NAME = "api.twitter.com"
-
-    def on
-      regist_to_shamrack
-      regist_to_shamrack(Net::HTTP.https_default_port)
-    end
-
-    def off
-      ShamRack.unmount_all
-    end
-
-    def on?
-      !ShamRack.application_for(API_HOST_NAME).nil?
-    end
-
-    private
-
-    def regist_to_shamrack(port = Net::HTTP.default_port)
-      ShamRack.at(API_HOST_NAME, port) do |env| 
-        if validate_request(env) && validate_user
-          generate_request 
-        end
-      end
-    end
-
-    def validate_request(env)
-      return false unless env && env['REQUEST_METHOD'] == 'GET' && env['PATH_INFO'] == '/'
-      # TODO 
-      # RequestHeaderの検証
-      true
-    end
-
-    def validate_user
-      # TODO
-      # ユーザ情報参照
-      # 事前登録したユーザとのマッチング
-      true
-    end
-
-    def generate_request
-      # TODO
-      # Request生成
-      status = '200 OK'
-      header = { 'Content-type' => 'text/plain' }
-      body = ['Hello, world!']
-      [status, header, body]
-    end
   end
 end
